+++ lib/collaboration_web/channels/topic_channel.ex
@@ -44,8 +44,22 @@ defmodule CollaborationWeb.TopicChannel do
     {:noreply, socket}
   end
 
+  def handle_in("create_idea", params, socket) do
+    params = Map.put params, "user_id", socket.assigns.user.id
+    case create_idea(params, socket.assigns.topic_id, socket.assigns.user) do
+      idea ->
+        idea = render_to_string IdeaView, "idea.html",
+          idea: idea,
+          user: socket.assigns.user
+
+        {:reply, {:ok, %{ idea: idea }}, socket}
+
+      _error ->
+        {:reply, :error, socket}
+    end
+  end
+
   def handle_in("create_comment", params, socket) do
-    IO.inspect params
     params = Map.put params, "user_id", socket.assigns.user.id
     case create_comment(params) do
       {:ok, comment} ->
@@ -93,3 +107,42 @@ defmodule CollaborationWeb.TopicChannel do
     end)
   end
 end
+
+# create automated feedback (if not admin or in condition 1 and 2
+# if !user.admin && user.condition in [3,4] do
+#   Task.Supervisor.async_nolink(Collaboration.TaskSupervisor, fn ->
+#     # wait for a little amount of time
+#     :timer.sleep Enum.random(5000..10000)
+#     # then create an automated feedback
+#     sequence = user.feedback_sequence
+#     feedback = case sequence do
+#       0 -> "1. Automated Feedback"
+#       1 -> "2. Automated Feedback"
+#       2 -> "3. Automated Feedback"
+#       3 -> "4. Automated Feedback"
+#       4 -> "5. Automated Feedback"
+#       5 -> "6. Automated Feedback"
+#       6 -> "7. Automated Feedback"
+#       7 -> "8. Automated Feedback"
+#       8 -> "9. Automated Feedback"
+#       9 -> "10. Automated Feedback"
+#     end
+#     # create automated feedback
+#     random_user = select_random_user(user.id)
+#     case create_comment(random_user, user, idea, %{text: feedback}) do
+#       {:ok, comment} ->
+#         # broadcast back to user
+#         Endpoint.broadcast("user:" <> user.id, "new_feedback", %{
+#           idea_id: idea.id,
+#           comment: render_to_string(CommentView, "comment.html",
+#             admin: user.admin,
+#             comment: load_comment(comment.id, user),
+#             user_id: user.id
+#           )
+#         })
+#     end
+#     # increase_feedback_sequence user
+#     sequence = if sequence === 9, do: 0, else: sequence + 1
+#     update_user(user, %{feedback_sequence: sequence})
+#   end)
+# end